refactor: extract receipt printer into printer.py
Move ESC/POS build and Windows/Linux print helpers out of app.py.
Keeps main behavior: always cp932, print_ticket returns True/False.
